# Remove support for spatial partitioning upon saving