--- a/roms_tools/tests/test_setup/test_boundary_forcing.py
+++ b/roms_tools/tests/test_setup/test_boundary_forcing.py
@@ -207,7 +207,7 @@ def test_boundary_forcing_save(boundary_forcing, tmp_path):
             str(tmp_path / file_str),
         ]:  # test for Path object and str
 
-            # Test saving without partitioning and grouping
+            # Test saving without grouping
             saved_filenames = boundary_forcing.save(filepath)
 
             filepath_str = str(Path(filepath).with_suffix(""))
@@ -217,7 +217,7 @@ def test_boundary_forcing_save(boundary_forcing, tmp_path):
             assert expected_filepath.exists()
             expected_filepath.unlink()
 
-            # Test saving without partitioning but with grouping
+            # Test saving with grouping
             saved_filenames = boundary_forcing.save(filepath, group=True)
 
             filepath_str = str(Path(filepath).with_suffix(""))
@@ -226,18 +226,6 @@ def test_boundary_forcing_save(boundary_forcing, tmp_path):
             assert saved_filenames == [expected_filepath]
             assert expected_filepath.exists()
             expected_filepath.unlink()
-
-            # Test saving with partitioning and grouping
-            saved_filenames = boundary_forcing.save(filepath, np_eta=2, group=True)
-            expected_filepath_list = [
-                Path(filepath_str + f"_202106.{index}.nc") for index in range(2)
-            ]
-
-            assert saved_filenames == expected_filepath_list
-
-            for expected_filepath in expected_filepath_list:
-                assert expected_filepath.exists()
-                expected_filepath.unlink()
 
 
 def test_bgc_boundary_forcing_plot(bgc_boundary_forcing_from_climatology):
@@ -282,20 +270,6 @@ def test_bgc_boundary_forcing_save(bgc_boundary_forcing_from_climatology, tmp_pa
             assert saved_filenames == [expected_filepath]
             assert expected_filepath.exists()
             expected_filepath.unlink()
-
-            # Test saving with partitioning
-            saved_filenames = bgc_boundary_forcing_from_climatology.save(
-                filepath, np_xi=2, group=True
-            )
-
-            expected_filepath_list = [
-                Path(filepath_str + f"_clim.{index}.nc") for index in range(2)
-            ]
-            assert saved_filenames == expected_filepath_list
-
-            for expected_filepath in expected_filepath_list:
-                assert expected_filepath.exists()
-                expected_filepath.unlink()
 
 
 @pytest.mark.parametrize(