--- a/roms_tools/tests/test_setup/test_initial_conditions.py
+++ b/roms_tools/tests/test_setup/test_initial_conditions.py
@@ -224,7 +224,6 @@ def test_initial_conditions_save(
             str(tmp_path / file_str),
         ]:  # test for Path object and str
 
-            # Test saving without partitioning
             saved_filenames = initial_conditions_with_bgc_from_climatology.save(
                 filepath
             )
@@ -234,20 +233,6 @@ def test_initial_conditions_save(
             assert filepath.exists()
             # Clean up the .nc file
             filepath.unlink()
-
-            # Test saving with partitioning
-            saved_filenames = initial_conditions_with_bgc_from_climatology.save(
-                filepath, np_eta=2
-            )
-
-            filepath_str = str(filepath.with_suffix(""))
-            expected_filepath_list = [
-                Path(filepath_str + f".{index}.nc") for index in range(2)
-            ]
-            assert saved_filenames == expected_filepath_list
-            for expected_filepath in expected_filepath_list:
-                assert expected_filepath.exists()
-                expected_filepath.unlink()
 
 
 def test_roundtrip_yaml(