## Changes

This PR removes the support for spatial partitioning upon saving—a functionality that, to my knowledge, is not actively used, not by C-Star at least. (Note that the user can still use ROMS-Tools to partition their *existing* files, via the `partition_netcdf` function.) The rationale of this PR is a performance gain, both in terms of faster compute times and memory savings. 

- [x] Tests removed that test partitioning upon saving
- [x] Passes `pre-commit run --all-files`
- [x] Changes are documented in `docs/releases.md`
- [x] Functions/methods are updated in `docs/api.rst`

## Issue resolved

To enable spatial partitioning we had to load the dataset into memory first before they were spatially partitioned and then saved to disk, see https://github.com/CWorthy-ocean/roms-tools/pull/114. This means we could not take advantage of Dask's optimized I/O capabilities. As of this PR, we never load the datasets into memory (when `use_dask = True`), they will simply be written to disk by Dask, which results in an improved memory footprint and faster computations.

## Example of performance gain

Saving 1 month of hourly surface forcing data for grid A in https://github.com/CWorthy-ocean/roms-tools/issues/201 with `use_dask = True` before and after this PR:

Before:
```
peak memory: 82923.36 MiB, increment: 82109.48 MiB
CPU times: user 7h 14min 36s, sys: 2min, total: 7h 16min 37s
Wall time: 8min 35s
```
After:
```
peak memory: 67603.79 MiB, increment: 66888.29 MiB
CPU times: user 6h 56min 22s, sys: 2min 20s, total: 6h 58min 42s
Wall time: 3min 37s
```
For context: Memory footprint of actual dataset: 23634.59 MB